Add endpoints to return counts of PPL expiries
There is a desire to have PPL expiries show in the performance metrics alongside revocations, but since these are done automatically without creating a task then they need a specific endpoint to return them.

Returns counts of PPL expiries within a date range and filtered to an establishment, grouped by schema version.
