+++ lib/api.js
@@ -3,6 +3,8 @@ const { NotFoundError } = require('@asl/service/errors');
 const errorHandler = require('@asl/service/lib/error-handler');
 const Knex = require('knex');
 
+const activeLicences = require('./routers/active-licences');
+const expiredProjects = require('./routers/expired-projects');
 const reports = require('./reports');
 
 module.exports = (settings) => {
@@ -18,47 +20,8 @@ module.exports = (settings) => {
 
   app.use('/reports', reports(settings));
 
-  app.get('/active-licences', (req, res, next) => {
-    Promise.resolve()
-      .then(() => {
-        const queries = ['projects', 'pils', 'training_pils', 'establishments']
-          .map(type => {
-            return req.db.asl(type)
-              .count()
-              .where({ status: 'active' })
-              .whereNull('deleted')
-              .where(builder => {
-                if (req.query.establishment) {
-                  const id = parseInt(req.query.establishment, 10);
-                  switch (type) {
-                    case 'pils':
-                    case 'projects':
-                      return builder.where({ establishment_id: id });
-                    case 'establishments':
-                      return builder.where({ id });
-                    case 'training_pils':
-                      return builder.whereExists(function () {
-                        return this.select('*').from('training_courses')
-                          .where({ establishment_id: id })
-                          .whereRaw('training_courses.id = training_pils.training_course_id');
-                      });
-                  }
-                }
-              })
-              .then(result => parseInt(result[0].count, 10));
-          });
-        return Promise.all(queries);
-      })
-      .then(([projects, pils, trainingPils, establishments]) => {
-        res.json({
-          projects,
-          pils,
-          trainingPils,
-          establishments
-        });
-      })
-      .catch(next);
-  });
+  app.use('/active-licences', activeLicences(settings));
+  app.use('/expired-projects', expiredProjects(settings));
 
   app.use(() => {
     throw new NotFoundError();
